requirements.txt
anyascii==0.3.3
asgiref==3.11.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
defusedxml==0.7.1
Django==5.2.9
django-cors-headers==4.9.0
django-filter==25.2
django-modelcluster==6.4.1
django-permissionedforms==0.1
django-stubs-ext==5.2.8
django-taggit==6.1.0
django-tasks==0.9.0
django-treebeard==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
greenlet==3.3.1
idna==3.11
joblib==1.3.0
laces==0.1.2
lightgbm==4.1.0
modelsearch==1.1.1
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.0
pillow==12.0.0
pillow_heif==1.1.1
playwright==1.58.0
psycopg2==2.9.11
pyee==13.0.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
RapidFuzz==3.14.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.4
telepath==0.3.1
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wagtail==7.2.1
wagtail-modeladmin==2.2.0
wheel==0.45.1
Willow==1.12.0
